chore: Move `service::auth::models` to a dedicated file
